requirements.txt
fastapi
python-multipart
uvicorn[standard]
google-genai
httpx
pydantic
python-dotenv
